+++ libft/ft_calloc.c
@@ -16,7 +16,7 @@
 void	*ft_calloc(size_t nmemb, size_t size)
 {
 	char	*ptr;
-	int		i;
+	size_t	i;
 
 	ptr = malloc(nmemb * size);
 	if (!ptr)
@@ -29,13 +29,3 @@ void	*ft_calloc(size_t nmemb, size_t size)
 	}
 	return ((void *)ptr);
 }
-
-// #include <stdio.h>
-// int	main(void)
-// {
-// 	char	*ptr = calloc(10, sizeof(int));
-// 	for (int i = 0; i < 10; i++)
-// 		printf("index %d has value %d \n", i, ptr[i]);
-// 	free(ptr);
-// 	return (0);
-// }
